Rename server package as example
