+++ codegen/example/example_cli_test.go
@@ -1,11 +1,11 @@
-package server
+package example
 
 import (
 	"bytes"
 	"testing"
 
 	"goa.design/goa/codegen"
-	"goa.design/goa/codegen/server/testdata"
+	"goa.design/goa/codegen/example/testdata"
 	"goa.design/goa/expr"
 )
 
@@ -26,7 +26,7 @@ func TestExampleCLIFiles(t *testing.T) {
 			// reset global variable
 			Servers = make(ServersData)
 			codegen.RunDSL(t, c.DSL)
-			fs := ExampleCLIFiles("", expr.Root)
+			fs := CLIFiles("", expr.Root)
 			if len(fs) == 0 {
 				t.Fatalf("got 0 files, expected 1")
 			}
