+++ codegen/example/server_data.go
@@ -1,4 +1,4 @@
-package server
+package example
 
 import (
 	"fmt"
