+++ codegen/example/testdata/dsls.go
