+++ codegen/example/testdata/example_cli_code.go
