--- codegen/server/docs.go
@@ -1,3 +1,0 @@
-// Package server contains code generation algorithms to produce an example
-// server and client implementation for the transports defined in the design.
-package server
