+++ http/codegen/example_cli_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"goa.design/goa/codegen"
-	"goa.design/goa/codegen/server"
-	ctestdata "goa.design/goa/codegen/server/testdata"
+	"goa.design/goa/codegen/example"
+	ctestdata "goa.design/goa/codegen/example/testdata"
 	"goa.design/goa/expr"
 	"goa.design/goa/http/codegen/testdata"
 )
@@ -26,7 +26,7 @@ func TestExampleCLIFiles(t *testing.T) {
 	for _, c := range cases {
 		t.Run(c.Name, func(t *testing.T) {
 			// reset global variable
-			server.Servers = make(server.ServersData)
+			example.Servers = make(example.ServersData)
 			codegen.RunDSL(t, c.DSL)
 			fs := ExampleCLIFiles("", expr.Root)
 			if len(fs) == 0 {
